Teste Unitário para Calcular a Taxa de Transferência

+++ src/main/java/br/com/fagner/enums/TransferSchedulingTypeStrategy.java
@@ -2,6 +2,6 @@ package br.com.fagner.enums;
 
 public interface TransferSchedulingTypeStrategy {
 	
-	public double calculateRate(int days, double valueTransfer);
+	public double calculateRate(long days, double valueTransfer);
 
 }
